fix: float culture size

+++ src/Entity/Cultures.php
@@ -109,12 +109,12 @@ class Cultures
         return $this;
     }
 
-    public function getSize(): ?int
+    public function getSize(): ?float
     {
         return $this->size;
     }
 
-    public function setSize(int $size): self
+    public function setSize(float $size): self
     {
         $this->size = $size;
